Revert "fixa upprepning av koden i ClassicGameView"
