+++ app/src/main/java/com/katja/proseccopong/ClassicGameView.kt
@@ -1,9 +1,10 @@
 package com.katja.proseccopong
 
 import android.content.Context
+import android.content.Intent
 import android.content.SharedPreferences
-import android.content.res.Resources
 import android.graphics.Canvas
+import android.graphics.Color
 import android.graphics.Paint
 import android.graphics.Rect
 import android.graphics.Typeface
@@ -11,16 +12,8 @@ import android.view.MotionEvent
 import android.view.SurfaceHolder
 import android.view.SurfaceView
 import androidx.core.content.ContextCompat
+import com.google.gson.Gson
 
-
-
-class ClassicGameView(context: Context, private val sharedPreferences: SharedPreferences,private val resources: Resources) : SurfaceView(context), SurfaceHolder.Callback,  GameView {
-    var mholder: SurfaceHolder? = null
-
-
-     var mcontext=context
-
-    private var platformLevel = 200f
 
 class ClassicGameView(context: Context, private val activityContext: Context, private val sharedPreferences: SharedPreferences) : SurfaceView(context), SurfaceHolder.Callback, Runnable, GameView {
     private var mholder: SurfaceHolder? = null
@@ -31,16 +24,17 @@ class ClassicGameView(context: Context, private val activityContext: Context, pr
     private var playerPlatform: PlayerPlatform
     private var thread: Thread? = null
     private var platformLevel = 100f
-
     private var platformHeight = 25f
     private var platformTop = platformHeight + platformLevel
     private var platformWidth = 200f
     lateinit var bounds: Rect
-
-
+    var viewWidth = 0f
+    var viewHeight = 0f
+    var paintPoints = Paint()
+    val textSizePoints: Float = resources.getDimension(R.dimen.text_size_points)
+    private var playerName: String = ""
     var existingScoreIndex = -1
-lateinit var gameManager: GameManager
-lateinit var proseccoGameView: ProseccoGameView
+
     init {
         mholder = holder
 
@@ -48,9 +42,9 @@ lateinit var proseccoGameView: ProseccoGameView
             holder?.addCallback(this)
 
         }
+        playerPlatform=PlayerPlatform(mcontext,platformWidth,platformHeight,0f,0f, platformLevel, Color.WHITE)
+        ball1 = Ball(this, mcontext,100f, 100f, 20f, 10f, 20f, platformTop)
 
-        gameManager= GameManager(this,context,"ClassicGame",sharedPreferences,resources)
-        proseccoGameView= ProseccoGameView(gameManager, mcontext,mcontext,sharedPreferences)
     }
 
     override fun surfaceCreated(holder: SurfaceHolder) {
@@ -59,18 +53,15 @@ lateinit var proseccoGameView: ProseccoGameView
 
     override fun surfaceChanged(holder: SurfaceHolder, format: Int, width: Int, height: Int) {
         bounds=Rect(0,0, width,height)
-     gameManager.   playerPlatform.initialize(width, height)
-      gameManager.  viewWidth = width.toFloat()
-      gameManager . viewHeight = height.toFloat()
-       gameManager. start()
+        playerPlatform.initialize(width, height)
+        viewWidth = width.toFloat()
+        viewHeight = height.toFloat()
+        start()
 
-    }
-    fun setPlayerName(name: String) {
-       gameManager. playerName = name
     }
 
     override fun surfaceDestroyed(holder: SurfaceHolder) {
-       gameManager. stop()
+        stop()
     }
     override fun onTouchEvent(event: MotionEvent?):Boolean{
         val touchX = event?.x ?: 0f
@@ -79,50 +70,153 @@ lateinit var proseccoGameView: ProseccoGameView
         val movementSpeed = 5f
 
         // Beräkna skillnaden mellan den nuvarande plattformens position och tryckpunkten
-        val difference = touchX -gameManager. playerPlatform.posX
+        val difference = touchX - playerPlatform.posX
 
         // Uppdatera plattformens position gradvis mot tryckpunkten
-        gameManager. playerPlatform.posX += difference / movementSpeed
+        playerPlatform.posX += difference / movementSpeed
 
         return true
     }
 
 
+    fun start() {
+        running = true
+        thread = Thread(this)
+        thread?.start()
+    }
 
+    fun stop() {
+        running = false
+        try {
+            thread?.join()
+        } catch (e:InterruptedException){
+            e.printStackTrace()
+        }
 
+    }
 
+    fun update() {
+        ball1.update()
+
+    }
 
     // Function to accsess the return statment of onIntersection by other classes
+    override fun ballDown(): Boolean {
+        return onIntersection(playerPlatform, ball1)
+    }
+
+    fun onIntersection(p:PlayerPlatform,b:Ball): Boolean{
+        // Calculate the centers of the platform and the ball
+        val platformCenterX = p.posX + p.width / 2
+        val ballCenterX = b.posX
+
+        // If the ball intersects the platform
+        if (b.posX - b.size / 2 <= platformCenterX + p.width / 2 && b.posX + b.size / 2 >= platformCenterX - p.width / 2) {
+            // Calculate the difference between the centers
+            val differenceX = ballCenterX - platformCenterX
+            // Reverse the ball's horizontal direction
+            b.speedX = differenceX / 2  // Adjust this factor as needed
+            // Reverse the ball's vertical direction
+            b.speedY *= -1
+            // Increse the ball's vertcal speed
+            b.speedY *= 1.05f // Adjust this factor as needed
+            // Move the ball up to avoid it going into the platform
+            b.posY = b.posY + b.speedY * 2
+            // Increment points
+            GameManager.addPoints()
+            return false // Return statment to mark that the ball is not out
+        }
+        else {
+            return true // Return statment to mark that the ball is out
+        }
+    }
+
+    fun draw() {
+
+        canvas= mholder!!.lockCanvas() ?: return
+        val backgroundDrawable = resources.getDrawable(R.drawable.black_background, null)
+        backgroundDrawable.setBounds(0, 0, canvas.width, canvas.height)
+        backgroundDrawable.draw(canvas)
+        drawPoints(canvas)
+        playerPlatform.draw(canvas)
+        ball1.draw(canvas)
+        holder!!.unlockCanvasAndPost(canvas)
+    }
+
+    override fun run() {
+        while (running) {
+
+            update()
+            draw()
+            ball1.checkbounders(bounds,mcontext)
+            playerPlatform.checkBounds(bounds)
+
+        }
+        Thread.sleep(6)
+    }
 
 
+    fun drawPoints(canvas: Canvas) {
+        // Rensa Canvas
+        val backgroundDrawable = resources.getDrawable(R.drawable.black_background, null)
+        backgroundDrawable.setBounds(0, 0, canvas.width, canvas.height)
+        backgroundDrawable.draw(canvas)
+
+        val textColor = ContextCompat.getColor(context, R.color.white)
+        val shadowColor = ContextCompat.getColor(context, R.color.gold)
+
+        paintPoints.color = textColor
+        paintPoints.textAlign = Paint.Align.CENTER
+        paintPoints.textSize = textSizePoints
+        paintPoints.typeface = Typeface.create(Typeface.DEFAULT, Typeface.BOLD)
+
+        // Använd shadowLayer för skugga
+        paintPoints.setShadowLayer(20f, 3f, 3f, shadowColor)
+
+        // Rita "Name" och "Score" bredvid varandra på samma rad, högre upp på skärmen
+        val nameText = "Name: $playerName".uppercase()
+        val scoreText = "Score: ${GameManager.points}".uppercase()
+        val printText = nameText + "     " + scoreText
+
+        val centerX = viewWidth / 2
+        val centerY = viewHeight / 10 // Justera y-koordinaten för att höja texten
+
+        canvas.drawText(printText, centerX, centerY, paintPoints)
+
+        paintPoints.clearShadowLayer()
+    }
 
 
-
-
-
-
-
-
-
-
+    fun setPlayerName(name: String) {
+        playerName = name
+    }
 
     // TODO: Ändra så att tidigare resultat inte skrivs över
 
+    fun saveScore() {
+        val editor = sharedPreferences.edit()
+
+        // Lägg till ny poäng i listan oavsett om det finns en duplicat
+        val newClassicScore = Score(playerName, GameManager.points, true)
+        ScoreList.scoreList.add(newClassicScore)
+
+        // Uppdatera den befintliga variabeln
+        existingScoreIndex = ScoreList.scoreList.indexOfFirst { it.name == playerName && it.classic }
+
+        // Konvertera ScoreList till en JSON-sträng och spara den i SharedPreferences
+        val scoreListJson = Gson().toJson(ScoreList.scoreList)
+        editor.putString("score_list", scoreListJson)
+        editor.apply()
+    }
 
 
 
-
-
-    override fun incrementPoints() {
-        TODO("Not yet implemented")
-=======
     override fun gameEnd(){
         saveScore() // Save the score before transitioning to HighscoreActivity
         println(ScoreList) //Sout for debug
         val intent = Intent(activityContext, HighscoreActivity::class.java)
         activityContext.startActivity(intent)
         GameManager.resetPoints() // Reset points variable so that it starts at 0 in the next game
-
     }
 }
