+++ app/src/main/java/com/katja/proseccopong/ClassicGameViewActivity.kt
@@ -10,8 +10,6 @@ class ClassicGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
 
     lateinit var binding: ActivityClassicGameViewBinding
     lateinit var classicGameView: ClassicGameView
-    lateinit var gameManager: GameManager
-    lateinit var proseccoGameView: ProseccoGameView
 
     // Skapa SharedPreferences-objekt
     private val sharedPreferences by lazy {
@@ -24,9 +22,7 @@ class ClassicGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
         setContentView(binding.root)
 
         val playerName = intent.getStringExtra("PLAYER_NAME")
-
-        classicGameView = ClassicGameView(this,  sharedPreferences,resources)
-        gameManager = GameManager(classicGameView, this, "ClassicGame", sharedPreferences, resources)
+        classicGameView = ClassicGameView(this, this, sharedPreferences)
         classicGameView.setPlayerName(playerName ?: "")
 
         binding.surfaceClassicGameView.holder.addCallback(classicGameView)
@@ -35,7 +31,7 @@ class ClassicGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
 
     override fun onDestroy() {
         super.onDestroy()
-        gameManager.saveScore() // Spara poängen när aktiviteten förstörs
+        classicGameView.saveScore() // Spara poängen när aktiviteten förstörs
     }
 
     override fun surfaceCreated(holder: SurfaceHolder) {
