+++ app/src/main/java/com/katja/proseccopong/ProseccoGameViewActivity.kt
@@ -10,8 +10,6 @@ class ProseccoGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
 
     lateinit var proseccoBinding: ActivityProseccoGameViewBinding
     lateinit var proseccoGameView: ProseccoGameView
-    lateinit var gameManager: GameManager
-    lateinit var classicGameView: ClassicGameView
 
     // Skapa SharedPreferences-objekt
     private val sharedPreferences by lazy {
@@ -24,10 +22,7 @@ class ProseccoGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
         setContentView(proseccoBinding.root)
 
         val playerName = intent.getStringExtra("PLAYER_NAME")
-
-        classicGameView= ClassicGameView(this,sharedPreferences,resources)
-        gameManager= GameManager(classicGameView,this,"ProseccoGame",sharedPreferences,resources)
-        proseccoGameView = ProseccoGameView(gameManager,this,this,  sharedPreferences)
+        proseccoGameView = ProseccoGameView(this, this, sharedPreferences)
         proseccoGameView.setPlayerName(playerName ?: "")
 
         proseccoBinding.surfaceProseccoGameView.holder.addCallback(proseccoGameView)
@@ -36,7 +31,7 @@ class ProseccoGameViewActivity : AppCompatActivity(), SurfaceHolder.Callback {
 
     override fun onDestroy() {
         super.onDestroy()
-        gameManager.saveScore() // Spara poängen när aktiviteten förstörs
+        proseccoGameView.saveScore() // Spara poängen när aktiviteten förstörs
     }
 
     override fun surfaceCreated(holder: SurfaceHolder) {
